+++ src/features/ens/components/profile/ProfileSheetConfigContext.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+export const ProfileSheetConfigContext = createContext({
+  enableZoomableImages: false,
+});
